backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
alembic>=1.13.0

# Auth
bcrypt>=4.0.0
pyjwt>=2.8.0

# YouTube pipeline — pin transcript API to last stable v0.x to avoid breaking v1.0 API change
yt-dlp>=2024.1.1
youtube-transcript-api==0.6.2